Fix section alignment
Not sure why, but when the alignment was 2 bytes, the GOT could end up offset by 2 bytes in the binary.

For example, I would compile the binary as though it was located at address 0x13370000. When I dump the binary I see the following:

```
133700a8 g       .got   00000000 _got_start

...

Contents of section .got:
 133700a8 1337005c ...
```

But when I do a hex dump of the binary I would see the following:

```
000000a0  53 80 66 fc 60 e4 13 37  00 5c ...
```

In this case, the GOT entry is 2 bytes forward from where it says it is.

Changing the alignment to 4 bytes and the GOT is still said to be located at address 0xa8, but when doing a hex dump of the binary I now see it at the correct address:

```
000000a0  53 80 66 fc 60 e4 00 00  13 37 00 5c ...
```